feat: rename module to cpulists on plural
