+++ src/cpulists/validate_test.go
@@ -1,4 +1,4 @@
-package cpulist
+package cpulists
 
 import "testing"
